Use lineWithFocusChart by default
We also enlarge charts to 200px to accommodate the new default
"lineWithFocusChart", keeping the other chart types to 150px.

Additionally, in order to resize the graphs properly to take into
account an expanded legend, we reworked the sizing for lineChart to
adjust the default margins to accommodate the legend under the
expanded height. Detailed comments are provided in the code.

Change-Id: I6eaeb70af32418971b1686c156d3e4bcfe6edb48