Remove "found uperf pid" after a test has completed
There is a "stop_server" function which echos a message
about the uperf pid found (and will kill). This message
was intended for "extra" uperf pids -before- a test,
to alert the user that old uperf pids were around.
This script is also used -after- a test, but the
messages are not necessary -we just want to kill
uperf pids after a test.  An option to echo or
not echo the message has been added to this function
so we don't get the messages after a test.