Get the bench-scrips unit-tests working
We had removed configtools from the pbench report and moved it to its
own repository, so now we just require it to be installed and available
for the unit tests.  As a result, we just needed to remove the in-tree
build of configtools performed by the unit test script.

And fix the pbench_uperf option processing which was emitting a debug
message for "--" instead of just recognizing the termination of the
options.