Further documentation cleanups
First, we have updated the README.md so that it can be used as input
to the `gh-pages` for the site.

We then updated the `doc/GETTING_STARTED` and `doc/INSTALL` to use
mark-down, refining them just a bit.

Finally we filed #70 to cover the `TODO` contents and removed it, and
then just removed the `test-rpms.org` file since it is no longer
applicable.