Remove random output of dbench unittests
The unit test for dbench will mostly pass, but fails every once in a
while due to the random order in which the dbench commands are run,
since it is putting them in the background and waiting for them all to
finish in parallel. For unit tests, it is sufficient to have
sequential behavior for now.

This is a bit of a hack, as we have to enable the dbench command with
knowledge of the presence of unit test execution.