Use updated merics for results
Pbench_uperf is the first to use updated metrics,
in order to standardize the reporting of iteration
results across benchmarks.  Four types of metrics
are avaulable:
Throughput
Latency
Resource
Efficiency

These are use in the iterations's result.txt, to be
processed by the benchmark script when generating
the summary result files.