Add mail notification for pbench-backup-tarballs
Make rsync produce some stats (in addition to the list of files)
and send the stats as notification mail.

Also exclude the TODO/TO-INDEX/etc directories from the backup: they
cause permissions problems (but I think the underlying reason is that
we have different UIDs for pbench on the two systems involved).

Fix unittests:

o Stub out mailx (which doesn't exist in the travis environment) and
rsync. Use the same method as in the agent/bench-scripts unit tests.

Invoke rsync through PATH in pbench-backup-tarballs in order to be
able to override it with the stub.

Add test-execution.log to unit test output.  Modify the original gold
files to add the (mostly empty) test-execution.log section.