Make sure that metadata.log is complete.
metadata-log is supposed to be called at the beginning of a run (with
a "beg" argument) and again at the end of the run (with an "end"
argument). Most bench scripts call it through collect-sysinfo.
user-benchmark is an exception: it calls it explicitly at the
beginning and through collect-sysinfo at the end. User scripts are a
problem (as usual).

Make metadata-log "end" check for the existence of the metadata.log
file and if not, make it call metadata_log_start explicitly to write
the file. Since the timestamp cannot be obtained from the date command
at that point, we make it use the $date variable that the base script
sets. That should make sure that we have a reasonable metadata.log
file if metadata-log "end" is called.

Since we use UTC for all other timestamps, make all timestamps be UTC
in the base script. That will have the user-visible side-effect of
tagging the results tarballs with the UTC timestamp instead of the
local time.

There are distressingly many instances of tarballs (mostly
user-benchmark) that don't have an end_run timestamp in
metadata.log. On the theory that that's caused by keyboard interrupts,
add a trap to user-benchmark that will catch INT and QUIT and run
metadata-log "end" to complete the metadata.log file. Also record the
fact that the run was interrupted in the metadata.log file.

Assuming that the above implementation is correct, plan on adding
the trap code to all the bench-scripts (TBD).