Make permisssions on unpacked files/directories more liberal.
Make pbench-unpack-tarballs chmod directories to at least 555
and files to at least 444.

Rationale: sometimes, results directories are moved into the pbench
results directory, before move-results is called to copy the whole
thing over. There have been cases where perms on those directories
were tight enough that the results were not visible on the web. The
above modes should be enough to alleviate that problem.

Fixed creation of $TMP/$hostname - this was removed from
pbench-base.sh and moved to this file in another branch, but the
deployed scripts were of mixed origin.