Add generic script to activate local installation.
Add script to copy the local config file and pbench keyfile
to their proper places.

The idea is that the generic pieces (pbench-agent and configtools)
will exist as packages based on upstream code.

The local bits (the local config file and the keyfile) will exist in
an internal package only: this package will have dependencies on the other
two, so when installed or updated, it will pull in those two if
necessary. Then it will stash the local bits in some place and call
this script in the post-install phase to copy them to their final
resting place.