Fix extra ";" when there are no tool options
Wen using register-tool, the tool options are
encoded in an array, and that array is built
with "$tool_opt_array_cmd" when registering a
remote tool.  We had a ";" separating the
tool_array_cmd and the register-tool command
but that is not needed when there are no tool
options.  So, now the ";" is removed from the
ssh call and only embedded in the tool_opt_array_cmd
when there are tool options