Fix perf consolidated report
The perf postprocess script will attempt to conslidate
sample counts for functions across different PIDs,
providing the user with a single entry for each
unique symbol and a total sample count.  We skipped
this feature when a callgraph was used, because the
report format was not friendly enough to postprocess.
The check for callgraph was not adequate, which has been
changed now.

If there is a easy way to generate a consolidated report
like this with "perf report", I would like to know.