Sort the list of tarballs by size before processing.
This is only a half-measure (or less): if there are multiple tarballs waiting
to be processed, it will process them in increasing size order, so
the short ones will finish sooner. But once a long one is encountered,
everything will be queued behind it, until it is finished. But it's a simple
change, already implemented and tested in pbench-index.

It also takes care of the (unlikely) event of long command lines.