Rename util scripts to avoid collisions.
All the util scripts now have a prefix of pbench-.

Move non-pbench scripts to contrib/ subdirectory.

contrib/ is not added to the PATH so there will be no
collision. If users want to use one or more of these,
they will need to call them by full pathname or add
contrib/ to their PATH.

Add a backward-compatibility wrapper temporarily.

We maintain compatibility with the old util-scripts names
*for one release only*. The old names will be symlinked to
this wrapper in a postinstall action when the RPM is installed.
The wrapper does some minimal checking, issues a big fat warning
and then executes the script under its new name.