Changes to facilitate installation of the pbench server.
The upstream code now provides pbench-activate-* scripts to a) copy
the local config files to where the other scripts expect them to be,
b) create a crontab file based on the pbench-server.cfg file and c)
create the structure that pbench-{move,copy}-results expects in the
document root of the HTTP server. It does *not* activate the crontab:
after the installation, a human is expected to give it the once-over
before unleashing it on an unsuspecting world.

This commit adds the activation scripts and cleans up the config files.
The example config files in the upstream repo are pretty faithful
representations of what we use internally, with only some email addresses
and server names sanitized.

We build a pbench-server RPM using the Fedora COPR infrastructure.

Internally, we also build a pbench-server-internal RPM that does the
following:

o it requires the upstream pbench-server RPM (and also the configtools
  RPM, also available on COPR).

o it packages the "internal" config files.

o in a post-install step, it calls the activation script.

Server activation is now split into multiple parts.  This is primarily to
facilitate installation without an "internal" package.

The steps are:

o Copy config files from where the internal package installs them to
  the final destination (by default /home/pbench/lib/config on the server).
  Without an internal package, this step consists of manually copying the example
  config files and editing them to taste.

o Create a crontab: assuming standard installation paths, this would be

  pbench-activate-create-crontab /home/pbench/lib/config /home/pbench/lib/crontab

  N.B. The crontab is *not* activated automatically in order to allow a human
  to take a look at it and make sure that nothing untoward has happened. After
  checking, the crontab can be activated (as user "pbench") with the command

  crontab /home/pbench/lib/crontab/crontab

o The server is assumed to be running apache. We need to create some files in
  apache's document root in order to allow move-results to copy results over.
  This is done (as root) with:

  pbench-activate-setup-results-host-info /home/pbench/lib/config

o pbench-activate-create-results-dir-structure creates the directory structure
  for the results. The cron scripts executed expect this structure to be in
  place.

Add script to create the required directories on the server and a combined script,
pbench-activate-server, that calls the rest of the activation scripts.