Add blazemeter taurus script to bench-scripts/contrib.
Added a description in the beginning of the file similar to the one found in the other bench tools.
Removed yaml (PyYaml) dependency. pbench_bzt now uses/generates json scenario files.
Added initial platform detection code.

Move pbench-bzt to bench-scripts/contrib: it will be moved
to bench-scripts/ (which is in the PATH) when it conforms
to the standard conventions. Right now, it's just a convenience
wrapper.