Fix travis build.
The travis build environment (ubuntu precise) supports python3.2
only. Recent pip apparently has *stopped* supporting python3.2.
We had to go through hoops to make the previous environment work,
but that is now broken. So we go through a different set of hoops
that make the build work, but there now seem to be unit test failures
that were not there before: they are under investigation.