Add support for specifying whether a jschart is a XY plot or timeseries
If timeseries is selected, the x-axis will use data formatting for
it's labels -- this also means that the x-axis datapoints need to be
valid date values.  Currently it has been tested with time since epoch
in millisecond values.

The XY plot format is the support that has always been there.  The
x-axis values are treated as raw values with no special meanings other
than their coordinate values.