Fix some bugs in the zoom_it function in jschart when the data model is timeseries
When the calculations are made in zoom_it, the date objects need to be
converted to integers prior to performing the calculations and then
converted back into date objects after the calculations are complete.

Additionally, the integers that are produced are extremely large.
This unveiled a secondary bug caused by the fact that the difference
in the extent/domain range was much smaller than the calculated zoom
change.  The calculations have been modified to use the range delta
for the calculations rather than the absolute values, so the size of
the values should no longer matter.