Add support for a new data model type called histogram
This new data model is very similar to the existing "xy" type as far
as the chart goes, but the table is handled much differently.  For a
histogram the data series values have a much different meaning than
traditional XY data series so the table has been updated to reflect
the interesting aspects of a histogram.