Safer kill: check for strange situations and deal with them.
Check that the pid corresponds to the tool before killing it.
If there is no process with that pid, eat the error.
If the kill succeeds, return success.
Otherwise check if the process is still running: kill it forcibly
if necessary, but return failure in that case.

pbench-kill-tools: Check that the list of pids to kill is not empty.