Modify the sar script (and its relatives) to handle user-defined options.
We now honor any options that are passed in when the tool is registered,
e.g. with

 pbench-register-tool --name=pidstat -- --options="-R -U"

However the postprocessing is skipped, since it is virtually certain
that the output is going to change and the postprocessing script is
going to break. It is expected that a user will provide his/her own
postprocessing in that case.