Allow different set-ups during server activation.
After installing the server, using the RPM from the COPR repository or
otherwise, the server setup has to be "activated": a config file has
to be copied in the proper place, the apache setup needs to be
modified, a data directory needs to be in place with the right
structure and a crontab needs to be generated.

This commit adds the activation scripts and cleans up the config files.
The example config files in the upstream repo are pretty faithful
representations of what we use internally, with only some email addresses
and server names sanitized.

Server activation is now split into multiple parts.  This is primarily to
facilitate installation without an "internal" package.

The steps are:

o Copy config files from where the internal package installs them to
  the final destination (by default <pbench-server-install-dir>/lib/config on the server).
  Without an internal package, this step consists of manually copying the example
  config files and editing them to taste.

o Create a crontab: assuming standard installation paths, this would be

  pbench-server-activate-create-crontab <pbench-server-install-dir>/lib/crontab

  N.B. The crontab is *not* activated automatically in order to allow a human
  to take a look at it and make sure that nothing untoward has happened. After
  checking, the crontab can be activated (as user "pbench") with the command

  crontab <pbench-server-install-dir>/lib/crontab/crontab

o The server is assumed to be running apache. We need to create some files in
  apache's document root in order to allow move-results to copy results over.
  This is done (as root) with:

  pbench-server-activate-setup-results-host-info

o pbench-server-activate-create-results-dir-structure creates the directory structure
  for the results. The cron scripts executed expect this structure to be in
  place.

The last three are called from pbench-server-activate which is what's
used in the posttrans section of the internal RPM.  CONFIG has to be
defined outside all the subordinate scripts: pbench-server-activate
defines it and exports it before calling them. The rest of the scripts
have to find what they need from the config file named by the CONFIG
env variable.

Fix section names and option names:

o There is no deploy section any longer; it is now pbench-server.

o mailto is in the DEFAULT section (i.e. in every section); by
  convention, we pick it up through the pbench-server section - there
  is no longer a mail section.