Allow activation of different setups.
Both agent and server now have "production" and "testing"
subdirectories in their respective config directories.  There are
scripts "pbench-{agent,server}-config-{activate,list}" that take
advantage of that structure and allow manual/automatic activation of a
particular setup.

The pbench-agent-local-activate script has been rewritten to install
the production configuration using the script above, instead of the
more ad-hoc method used previously. It still installs the ssh private
key the same way as before.

The upstream code now provides pbench-activate-* scripts to a) copy
the local config files to where the other scripts expect them to be,
b) create a crontab file based on the pbench-server.cfg file and c)
create the structure that pbench-{move,copy}-results expects in the
document root of the HTTP server. It does *not* activate the crontab:
after the installation, a human is expected to give it the once-over
before unleashing it on an unsuspecting world.

This commit adds the activation scripts and cleans up the config files.
The example config files in the upstream repo are pretty faithful
representations of what we use internally, with only some email addresses
and server names sanitized.

Server activation is now split into multiple parts.  This is primarily to
facilitate installation without an "internal" package.

The steps are:

o Copy config files from where the internal package installs them to
  the final destination (by default /home/pbench/lib/config on the server).
  Without an internal package, this step consists of manually copying the example
  config files and editing them to taste.

o Create a crontab: assuming standard installation paths, this would be

  pbench-server-activate-create-crontab /home/pbench/lib/crontab

  N.B. The crontab is *not* activated automatically in order to allow a human
  to take a look at it and make sure that nothing untoward has happened. After
  checking, the crontab can be activated (as user "pbench") with the command

  crontab /home/pbench/lib/crontab/crontab

o The server is assumed to be running apache. We need to create some files in
  apache's document root in order to allow move-results to copy results over.
  This is done (as root) with:

  pbench-server-activate-setup-results-host-info

o pbench-server-activate-create-results-dir-structure creates the directory structure
  for the results. The cron scripts executed expect this structure to be in
  place.

The last three are called from pbench-server-activate which is what's
used in the posttrans section of the internal RPM.  CONFIG has to be
defined outside all the subordinate scripts: pbench-server-activate
defines it and exports it before calling them. They have to find what
they need from the CONFIG env variable.

Fix unittests.

Fix section names and option names:

o There is no deploy section any longer; it is now pbench-server.

o mailto is in the DEFAULT section (i.e. in every section); by
  convention, we pick it up through the pbench-server section - there
  is no longer a mail section.