Benchmark and iteration summary/processing scripts
Significant changes in support of using common
generate-benchmark-summary and process-iteration-samples scripts.
This eliminates a lot of code in pbench-uperf.

The overall goal is to enable a facility for generating well formed
metadata for benchmark results to enhance what we index into the
ElasticSearch warehouse today.

The new generate-benchmark-summary script produces JSON data as well
as new txt and html summary reports.

The new process-iteration-samples script will be used by all benchmarks
which produce a pbench-standard JSON data for their metrics.
This script will compute metrics like average and standard deviation,
and will also indicate to the benchmark script if the iteration
"passed" based on maximum allowed stddev-percent.

Add .cmd files for debugging later.  The command files are for the
pbench-uperf script, the process-iteration-samples script, and the
uperf-postprocess script.  This allows a user to re-run any of those
scripts after a test to re-do post-processing.  Note that "pbench-uperf"
command still requires the --postprocess-only=y and --run-dir=
to work.

Every benchmark [should] have multiple test samples for any
benchmark iteration.  The samples for the iteration must be
examined to compute the mean and standard deviation, and
do some houskeeping, like tar'ing up sample data, and
letting the caller know if this iteration passed the max-stddev
threshold.

All benchmarks which use the new JSON metric format can use
process-iteration-samples.  For unittests, all we need are:
1: multiple sample directories, each with a result.json
2: a process-iteration-samples-args file so the unittest
   uses the right arguments (often benchmark specific)
That's it.  As each benchmark adopts the JSON format,
a new samples/process-iteration-samples-N directory
should be created, with multiple benchmark iteration
sample subdirs, each with result.json files for that
benchmark.  The current one is for uperf.

Moved aggregate-throughput and efficiency metric calculation from
uperf-postprocess to BenchPostprocess so all benchmarks can benefit
from these functions. Both now require JSON formatted %workload hash.

By default we keep non-reference and failed iteration data. The
"reference-result" is only for throughput, but users may want to see
reference result for latency, or efficiency, which may be a different
sample.

Keeping failed tool data by default helps us figure out why stddev
could not be met.