Add verify script for backup tarballs.
Do not assume that mounts exist on the primary and backup archive
locations.  That way, we can run the verify script even if we are in
maintenance mode and one or the other (or both) of the directories is
missing.

Change both backup scripts to follow the same convention and change
the config file to issue the arguments in the correct order for these
two scripts:

  - arg1: backup directory
  - arg2: pbench deploy directory
  - arg2: pbench deploy directory (local)

The archive directory is calculated in pbench-base.sh.

Typical invocation:

   pbench-{backup,verify}-tarballs /pbench-local/archive.backup /pbench /pbench-local

Unit tests for pbench-verify-backup-tarballs

9.1: normal case - archive and backup agree - md5s are fine.
9.2: more tarballs in archive than in backup
9.3: more tarballs in backup than in archive
9.4: md5 error in archive
9.5: md5 error in backup