Use pidof -x to get list of pids.
Each tool script belongs to a 3- or 4-deep process hierarchy of which
pidof returned one pid (and not necessarily the one that was stored in
the tool pid file); consequently, safe_kill would skip it and there were
tools (and their screens) hanging around at the end.

Use pidof -x to get the complete list.

Change the relevant unit tests (and the mocked pidof) to accommodate
this change.