pbench-list-triggers rewrite.
Simplify: get rid of --name which seems applicable to pbench-list-tools
but not to pbench-list-triggers - triggers are associated with groups.

There is a single tool-triggers file that is created by pbench-register-tool-trigger,
so simplify the processing accordingly.

Add some unit tests: the pbench-tool-trigger-test script is only used for
unit testing but because of limitations in the unittests script, it has to
live in the util-scripts directory. This limitation will be lifter in the
future and scripts like pbench-tool-trigger-test will be moved to their
own subdirectory.