Templating prototype for `pbench-fio`, using config files to specify fio parameters. The order of precedence used is as follows: (from high to low)
- Command line options already given to pbench-fio
- Fio job file given with --job-file

The defaults in agent/bench-scripts/templates/fio.job match the existing
defaults Additionally the template currently has the following two special
semantics:

- '$target' gets replaced with each of the specified targets, ultimately
making a copy of each of the job file sections with '$target' in
their name for each target.

- '$@' gets replaced with whatever was given on the command line of
make-fio-jobfile.py. For now this only works for 'bs' and 'rw', so that
I didn't have to make substantial changes to pbench-fio - namely
reconstruct the for-loops which iterate over the targets, test_types,
and block_sizes.

make-fio-jobfile.py handles both python 2.7 and 3, and should handle
with and without latest configparser installed.