Sync satellite to master.
Add two scripts to sync a satellite to the master server.

Usage: .../pbench-rsync-satellite <prefix> <remotehost> <remotearchive> <TOP> [<TOP_LOCAL}]

The script visits the remote archive on the remote host, gathers up a
list of controllers that have submitted results and, for each
controller, rsyncs all the tarballs and their md5s to the archive on
the master server.

[TBD: It then checks the tarballs against their md5 and deletes all
that pass from the remote archive.]

It then creates a work queue of files and sizes, sorted in increasing
order by size and calls the second script with the work queue as
argument.

The second script schedules the tarballs for unpacking: it selects a
prefix of the work queue whose total size is less than 50MB and
creates links in the TODO directory for each one. It then waits for
the last link on that list to disappear.  That happens when
pbench-unpack-tarball finishes with it. Any files that are bigger than
50MB are submitted individually. If the scheduler has to wait more
than 30 mins for a set to finish, it exits saving the list of files
still to be processed.  In that case, mail is sent to alert a human.

The idea is to limit the number and size of tarballs so that we don't
overtax Gluster and does not delay other tarballs unduly. The 50MB max
size seems to limit processing time to less than 10-12 mins but it's
not clear what will happen with much larger tarballs.