Change the wildcard path to fiologparser_hist.py
fiologparser_hist.py was complaining that it could not
find the clat_hist files it was looking for.

That's because fio --client pulls back the *_clat_hist.*.log.* files
under the sampleX/ directory, not under sampleX/clients/<mumble>.

Also added a couple of fixes to eliminate noise: made a stdout
redirection into a stderr redirection and added a -f to rm.