Index csv tool files
Get the iterations out of the metadata log, and loop over
iterations/samples/hosts/tools/csv files, converting
the csv file into a dict that can be added to the indexing
actions.

This turns out to be too slow: more than 30 mins on perf42
for a relatively small, 10-iteration uperf run. The culprit
is the extractfile() functionality of tarfile.py, which scans
through the whole tarball every time, looking for the last
instance of each file (there is only one instance, but tarfile.py
has to check in any case).

To work around this, make pbench-index set up a temp directory and let
index-pbench use it to unpack the tarball once and for all
initially. It then accesses files in the temp directory through the
file system, instead of using extractfile(). That cuts down the
processing time by a factor of 10 or so.