pbench-metadata-log: reduction of ssh invocations
Copy remote tools-$group directories locally and use the local
copies to build the metadata log entries.

The local copies are done to a tmp directory under the current
run directory of the benchmark, which requires all the bench scripts
to export benchmark_run_dir (pbench-user-benchmark was the only one
that was doing that before: see issue #349 and PR #351).

Modify all the benchmarks toexport benchmark_run_dir, so we use it to
set the tmpdir underneath it: we don't use the global tmp dir to avoid
stale entries.

Add separate test-bin directory for util-scripts unittests:
it used to be a symlink to the bench-scripts/test-bin directory.
It is now a bona-fide directory containing mostly links to the
corresponding files in ../bench-scripts/test-bin, except for "scp"
which overrides the one under bench-scripts.

util-scripts/test-bin/scp now pretends to copy tools-default
directories from the remotes, so that we can check the newly
introduced for-loop in pbench-metadata-log. Fix test-03 for
the new output.

Issue #485