Pbench agent base script: do not recalculate $date
$date is set by the base script which is sourced by every benchmark
script, tool script and util script. It is used by benchmark scripts
to create the results directory for the run, but because it is
recalculated every time the base script is sourced, other components
were getting a different value and e.g. could not find the results
directory for the run.

We now check whether $date has a value already and if so, we just
leave it alone. It is only set the first time that the base script
is sourced.