Fix unit test
Fix the date in the metadata.log of the tarball for test-7.8 to make
it agree with the date in the name of the result directory. This is
what would have happened if we had the "pbench agent base script: do
not recalculate $date" fix when we produced the tarball.

That also changes the MD5 sum and consequently the id and the parents
in the JSON document.