Use toolgroup; check column header subfield
We were erroneously using the default tool group regardless of the
record run's tool group. We now use the proper tool group recorded
with the result.

We now verify that a subfield found in a column header is one of the
expected subfields as found in the handler data description.