[wip] First pass at iostat tool data handling
This work addresses issue #489. We have implemented handling of iostat
.csv files in a data-driven form, allowing us to add support for other
tools via a new data set that describes that tool. As this is a first
pass, it is likely we'll need to refine this, and add support for older
data sets from past pbench-agent versions.

We have performed a radical refactoring of the index-pbench code base
to enable this:

  1. We adding a PbenchTarBall class to encapsulate all the parameters
     being passed around

  2. We coalesced the code related to the three different objects
     indexed into one location, and performed some refactoring to
     reduce multiple levels

  3. We pushed more of the ES related code down into routines to add
     clarity

Finally, for index-pbench, we changed to emit properly formatted JSON
rather than just using pprint.