Final changes for automating server installation
With these changes, server installation requires
installing two RPMs (pbench-server-internal and pbench-web-server)
and two manual steps:

 - Install public key into pbench users authorized_keys file.
 - Enable the generated crontab after inspection.

Changes:

Uncomment the chown of $pbench_dir:
Not sure why that was commented out, but it is necessary.

Add -Z option to mv:
When moving the results from the tmp directory to incoming/, with
selinux enabled, the files need to be relabeled as httpd_content_t
in order to allow httpd to access them. `mv -Z' does that and it does
not create any problems on non-selinux-enabled systems.

Add script to start httpd:
Add a script to start httpd and make the appropriate links into the
document root. Call the script from the postinstall
pbench-server-activate script.

Add script to make the static link for pbench-web-server. Add this
script to the pbench-web-server postinstall actions.

Make script executable and fix unit tests:
Make pbench-server-activate-start-httpd executable.
Mock out systemctl and ln for use in unit tests.
Make the ln in unittests/_setup_state be the "real" ln (/bin/ln).
Fix the test-8 gold output.