Fix the timestamp format
The format is YYYY-MM-DD and is now shared between the timestamp
function and the date variable in agent/base.

More importantly, this makes pbench-metadata-log use that same format
for the start_run and end_run fields: the indexer assumes this format,
but we will fix up these fields if necessary in the back-end as a
stop-gap.