Change time format in suffix of run directory
Currently the time format used in the directory names for the benchmark
run directory is:

    TS_FORMAT="%FT%H:%M:%S"

We'd like to change it to:

    TS_FORMAT="%Y.%m.%dT%H.%M.%S"

Making this change allows us to pass the benchmark run directory to
docker to bind mount into a container.

Currently, when trying to export $benchmark_run_dir inside container,
E.g.:

    docker run --privileged -it -v $benchmark_run_dir:/pbench image bash

it will fail with:

    /usr/bin/docker-current: Error response from daemon: Invalid bind
    mount spec "/var/lib/pbench-agent/pbench-user-benchmark_sysbench"
    "incontainer_2017-04-11_15:17:06:/pbench": Invalid volume specif"
    "ication: '/var...

The colon in the directory name is not handled properly by docker as,
that is interpreted by docker as point to look for source:destination
when "-v" volume flag is used.

Signed-off-by: Elvir Kuric <elvirkuric@gmail.com>