Update the rsync scrip to remove slowness and some other fix
The trouble is that the rsync is run for every host, whether that host has any tarballs for copying or not. Unfortunately, that takes about 5 seconds, even if the host directory is empty. For a couple of hundred hosts, we end up spending 20 minutes in this loop, even if there is nothing to be done.

But now the script is updated to sync only that hosts, which has tarballs to copy. Two more state directories are made, named TO-SYNC and SYNCED, to track which tarballs are getting synced. If a user, uses any prefix while moving result it will saved under .prefix/ in the production server. It will also get deleted from the .prefix/ on the remote server.

Update to create TO-SYNC and SYNCED state directories

1. pbench-base.sh also needs changes to create TO-SYNC and SYNCED state subdirs.
2. Changes made to avoid listing the subdirs that are excluded from the rsync explicitly.

Include python script to cleanup

It is a good idea to cleanup the tar, md5 and prefix file with a
single python script instead of multiple ssh.

Update to make it work with executable file