Add disk tool to pbench
Disk tool gives us the information about the disk space usage,
free space, filesystems that have been mounted.

A benchmark run with the disk tool registered will generate csv
files with the data, a html file for rendering graphs which can
be viewed when the results are moved to the web server.