Add a way to quarantine pbench tarballs that don't match their MD5 sums
XXX - for now backup supposed to be in the same server as archive

Update the script to do the followings:
1. Tarballs in backup that fail the MD5 check will be moved to quarantine directory
(assuming that is under backup itself)
2. Tarballs in archive that fail the MD5 check will be restored from backup
(assuming that the backup is good).
3. Tarballs that fail the MD5 check in backup as well as in archive will be moved to
quarantine directory(assuming that is under archive itself)

Moving tarballs means it will move tar files as well as md5 files. In case of archive,
tar, md5 and prefix all files will be moved to quarantine directory.