pbench-rsync-satellite: bug fixes and unittest prep
We execute much more of the script now, so we have to
distinguish between unit test runs and "real" runs. For
unit test runs, we don't use the pid for temp directories
and we don't schedule any tarballs for unpacking.

Trying to get the test for bad MD5 to pass, uncovered a couple
of bugs in the format of various files that are used to record
such errors.