pbench-rsync-satellite: add unit test
Add a unit test to force mail to be sent. That required breaking
the MD5 of a tarball and allowing the script to go through almost
its complete flow (there are still mock rsync and ssh, but md5sum
is no longer mocked).

Fixed a reproducibility problem (using  a pid for unique temp
directories does not work for unit tests) that required a fix
to test-10 as well.

Fix wrong redirection for pushd (it was printing out a local directory
which broke the unit test) and fix the gold output.