Fix fio-postprocess
Closedir was called unconditionally and failing if the opendir had failed.
Move closedir inside the if that opens the directory successfully.

Make warning more explicit by including the problematic filename.

Refine duplicate time stamp test:

fio_[c]lat.1.log contains entries of the form

timestamp, value, rw_type, ??

where rw_type = 0 for read, 1 for write.  The fourth field seems to be
0 always, but that's not guaranteed. In any case, it is not used in
fio postprocessing, so we ignore it.

fio-postprocess was complaining about duplicate time stamps, but
duplicate timestamps *for different rw_types* are allowed. So change
the test in fio-postprocess to check both.

Changes:
   BenchPostprocess.pm: add rw_label; change value_exists() to return
      the index where the value of the given field was found (-1 on
      failure), instead of a boolean 1/0.

   fio-postprocess: check if the timestamp value exists, but also
      whether the rw_type at that index is the same as the new
      sample. If so, complain, but if not add the sample.

There is still a duplicate time stamp at the very end of these two files:
apparently this corresponds to the final partial operation that fio
was doing. It's unclear why the timestamp is duplicated, but it is -
the final resolution might be to get upstream fio to not do that.
For the time being, we live with the warning.

Perl was issuing a warning: "push on reference is experimental at line 230".
Eliminate warning.