Do not assume that unit tests are sequentially numbered
The unit test names are the indices of the cmds/tools associative
array. If a unittests script is called without arguments, instead of
assuming that the tests are numbered within a sequential range, derive
the complete set of names from the indices of the array and sort them
(otherwise, they appear in hashtable ("random") order).

Modified bench-scripts and utils-scripts unittests accordingly.  The
server unittests script already did that, the tools/postprocess
unittests script uses whatever it finds in the samples subdirectory,
so it does not need this treatment.