move, clear results only when set in vars
pbench-agent container runs pbench-clear-results and pbench-move-results
commands before and after each benchmark run. This commit fixes it to
run the commands only when the move_results and clear_results vars are
set to true in the vars file. This will allow us to move, clear the
results only when we want to.