Add sync-package-tarballs script
This script will collect all the tarballs, its md5 and its prefix if present,
from different hosts and package it in a single tarball and return it to the
server for futher processing. This will be a single file so it is easy to sync
as we don't have to run rsync on each host.