Update the script to unpack tarballs and create symlinks for them
This is a new unpack script which will decide the unpacking directory
according to the path it gets as a parameter. If this is script is
running in any other server other than production, it will unpack in
the tmp directory (considering that the path given in parameter is tmp)
and create a symlink of the content in the incoming as well as in the
results directory.