Add results indexing
Look for result.json files in the tarball and index them.

Currently, the only transformation applied is to change
the last entry of a timeseries so that the value is a float:
in some cases, it was (integer) 0 in and because the timeseries
is an array, ES raised an error (incompatible types).

Eventually, we'll get rid of all arrays and "denormalize" the entries.

Not every bench script produces result.json files. Deal with
that situation gracefully.