uperf-postprocess: force port_labels to be strings
Depending on the perl version, these port labels might
be represented as intergers or strings.  Aggregate
metrics also use "all" as the port label, so let's just
force all port labels to be strings.  Port labels are
not processed as numbers anyway, even though they happen
to be a number (in most cases).  Forcing strings also
ensures the different perl versions should always produce
strings for this data when converting to JSON.